+++ abics/scripts/main_dft_latgas.py
@@ -26,15 +26,11 @@ import scipy.constants as constants
 
 from abics import __version__
 from abics.mc import CanonicalMonteCarlo, RandomSampling
-from abics.mc_mpi import (
-    RX_MPI_init,
-    TemperatureRX_MPI,
-    PopulationAnnealing,
-    RXParams,
-    PAMCParams,
-    ParallelRandomParams,
-    EmbarrassinglyParallelSampling,
-)
+
+from abics.sampling.mc_mpi import RX_MPI_init
+from abics.sampling.rxmc import TemperatureRX_MPI, RXParams
+from abics.sampling.pamc import PopulationAnnealing, PAMCParams
+from abics.sampling.simple_parallel import EmbarrassinglyParallelSampling, ParallelRandomParams
 
 from abics.applications.latgas_abinitio_interface import (
     DefaultObserver,
